Fix cargo fmt formatting issues
Auto-format long lines in cli/args.rs, snapshot.rs, and fc-agent/main.rs

+++ src/commands/common.rs
@@ -889,7 +889,7 @@ pub async fn create_snapshot_core(
             memory_bytes
         };
         if let Ok(stat) = nix::sys::statvfs::statvfs(snapshot_dir) {
-            let available_bytes = stat.blocks_available() * stat.fragment_size() as u64;
+            let available_bytes = stat.blocks_available() * stat.fragment_size();
             if available_bytes < required_bytes {
                 anyhow::bail!(
                     "Not enough disk space for {} snapshot: need {} MiB, have {} MiB free on {}. \
